Fixed salinity parsing bug Improved salinity display

+++ src/info/curtbinder/reefangel/phone/XMLHandler.java
@@ -184,7 +184,7 @@ public class XMLHandler extends DefaultHandler {
 		} else if ( tag.equals( Globals.xmlPWMDaylight ) ) {
 			ra.setPwmD( Byte.parseByte( currentElementText ) );
 		} else if ( tag.equals( Globals.xmlSalinity ) ) {
-			ra.setSalinity( Byte.parseByte( currentElementText ) );
+			ra.setSalinity( Integer.parseInt( currentElementText ) );
 		} else if ( tag.equals( Globals.xmlRelay ) ) {
 			ra.setMainRelayData( Short.parseShort( currentElementText ) );
 		} else if ( tag.equals( Globals.xmlRelayMaskOn ) ) {
